parental-control-backend/services/p-gateway-simulator/requirements.txt
# Core dependencies
confluent-kafka==2.3.0
PyYAML==6.0.1
python-dateutil==2.8.2

# AWS SDK
boto3==1.34.19
botocore==1.34.19

# Utilities
python-json-logger==2.0.7
prometheus-client==0.19.0

# Type hints and validation
pydantic==2.5.3
typing-extensions==4.9.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
